Massive additions in testing and algorithms:
  - Added new inner classes to NearestNeighborHeapSearch
  - Better tiered abstraction in clustering algorithm classes
  - Refactored some classes into other packages
  - Re-designed NearestNeighbors and added RadiusNeighbors
  - Added NearestCentroid
  - Created UnsupervisedIndexAffinity algorithm for unsupervised scoring
  - Restructured unit testing suite, achieved 85% coverage

Next steps:

  - Complete HDBSCAN
  - Test BoruvkaAlgorithm
  - Redesign KMeans and KMedoids algorithms for better accuracy and efficiency
  - Improve class and method documentation

+++ src/com/clust4j/utils/DistanceMetric.java
@@ -1,7 +1,7 @@
 package com.clust4j.utils;
 
 public interface DistanceMetric extends GeometricallySeparable {
-	public static final int DEFAULT_P = 2;
+	public static final double DEFAULT_P = 2.0;
 	
 	public double getP();
